Refactor Duplicate interfaces/type
