+++ src/hooks/useGames.ts
@@ -2,12 +2,9 @@ import {  useQuery } from "@tanstack/react-query";
 import { GameQuery } from "../App";
 import { FetchResponse } from "../services/api-client";
 import apiClient from "../services/api-client";
+import { Platform } from "./usePlatforms";
 
-export type Platform = {
-  id: number;
-  name: string;
-  slug: string;
-};
+
 
 export type Game = {
   id: number;
